Adds a 'NotProvided' solution state to the SolutionState enum (#29)
This is not well documented but one of the values in the SolutionState bit at 0x32 is 0x0002, which indicates that the solution is not present in the file. In practice I think almost every client treats it just about the same as 0x0004 in that spot (suppressing or disabling "reveal" tools) but it may be the case that some clients will correctly not offer an attempt to unscramble either.